requirements.txt
langchain-openai==0.0.5
python-dotenv==1.0.0
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1 